requirements.txt
fastapi==0.110.0
uvicorn==0.30.1
pydantic==2.7.1
python-dotenv==1.0.1
openai==1.12.0
tiktoken==0.6.0
beautifulsoup4==4.13.1
requests==2.31.0
psutil==5.9.8
python-multipart==0.0.9
rich==13.7.1
gunicorn==21.2.0
numpy==1.23.5
librosa==0.10.1
scipy==1.10.1
pydub==0.25.1
SpeechRecognition==3.10.1
transformers==4.30.2
nltk==3.8.1
pyAudioAnalysis==0.3.14
opensmile==2.4.2
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
tokenizers==0.13.3
crepe==0.0.13
